UBUNTU: SAUCE: XEN: Do not force sched_clock_stable
BugLink: http://bugs.launchpad.net/bugs/727459

This is part of the changes that have been found while syncing up
the patchsets. Apparently when applying the endif went into the
wrong place and causes the clock to be assumed always stable. Which
causes problems on certain CPUs (probably the fork/clone hang as well
as processes getting stuck for a longer time).

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Acked-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>