cve-tools: complete rewrite to a streamed approach
Complete rewrite of the cves-sync tool.  This is now a streamed approach
which means we only need to scan each of the git logs once improving
performance about 10x.  We also have a greater association with tags
which makes reporting more accurate where there are security releases
merged into master.

Signed-off-by: Andy Whitcroft <apw@canonical.com>