Revert "USB: xhci - fix math in xhci_get_endpoint_interval()"
This reverts commit fe3bf5f4be7d3a1a61225c77bd8a9b51baa3ac19.

Due to a reported problem with USB3 after application of three patches
from upstream stable, these patches are being reverted until they can
be tested seperately and a determination made of whether there is a
problem and how bad it is. The reporter has not yet responded to a
request to test kernels with each of the most suspect patches reverted,
so all will be reverted. Testing will proceed using these patches and
hopefully they can be restored soon. The USB3 problem was reported
in LP bug number 769042